+++ Picard.h
@@ -3,9 +3,9 @@
 
 // These Picard's method iteration functions calcs solution for ODE: u'(x) = x^2 + u^2
 
-double PicardFirstIteration(double x);
-double PartialPicardSecondIteration(double x);
-double PartialPicardThirdIteration(double x);
+long double PicardFirstIteration(long double x);
+long double PartialPicardSecondIteration(long double x);
+long double PartialPicardThirdIteration(long double x);
 
 #endif // PICARD
